requirements.txt
atomicwrites==1.4.0
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
idna==3.3
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
PyJWT==2.3.0
pyparsing==3.0.7
pytest==7.0.1
python-dotenv @ https://github.com/theskumar/python-dotenv/archive/master.zippytz==2021.3
requests==2.27.1
tomli==2.0.1
twilio==7.6.0
urllib3==1.26.8